Move script api into separate file
