+++ include/mainwindow.h
@@ -38,6 +38,7 @@ public:
     MainWindow(const MainWindow &) = delete;
     MainWindow & operator = (const MainWindow &) = delete;
 
+    // Scripting API
     Q_INVOKABLE QJSValue getBlock(int x, int y);
     Q_INVOKABLE void setBlock(int x, int y, int tile, int collision, int elevation);
     Q_INVOKABLE void setBlocksFromSelection(int x, int y);
